+++ includes/parsing.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: oaarsse <oaarsse@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ple-stra <ple-stra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/30 15:59:06 by ple-stra          #+#    #+#             */
-/*   Updated: 2022/10/11 13:35:32 by oaarsse          ###   ########.fr       */
+/*   Updated: 2022/10/10 19:37:44 by ple-stra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,8 +36,7 @@ alone, please give them a friend"
 # define ERR_CSM_MULTIPLE_OUTPUTS "%s: warning: Multiple outputs are given, \
 \"%s\" will be ignored."
 
-t_command		*get_command(char *line);
-t_command		*parsing(t_prg_data *prg_data, char *line);
-t_bool			parsing_validation(char *line);
+t_command		**get_commands(char *line);
+t_command		**parsing(t_prg_data *prg_data, char *line);
 
 #endif
