+++ srcs/parsing/parsing.c
@@ -6,25 +6,41 @@
 /*   By: oaarsse <oaarsse@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/03 19:36:38 by oaarsse           #+#    #+#             */
-/*   Updated: 2022/10/11 13:36:06 by oaarsse          ###   ########.fr       */
+/*   Updated: 2022/10/07 15:10:27 by oaarsse          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "common.h"
 #include "parsing.h"
 
+// Checks for empty line
+static int	is_empty(char *line)
+{
+	int	i;
+
+	i = 0;
+	if (!line)
+		return (1);
+	while (line[i])
+	{
+		if (!ft_isspace(line[i])) // TODO: check if all whitespaces are ignored in the command input
+			return (0);
+		i++;
+	}
+	return (1);
+}
+
 // Takes as input the raw user input from the shell
 // Will fill the program data struct with the commands to exec
 // Also checks if the line is not empty
-t_command	*parsing(t_prg_data *prg_data, char *line)
+t_command	**parsing(t_prg_data *prg_data, char *line)
 {
 	(void)prg_data;
-	if (!line)
+	if (is_empty(line))
 		return (NULL);
-	while (*line == ' ' || *line == '\t') // TODO: check if all whitespaces are ignored in the command input or just space and tab
+	// TODO: check if all whitespaces are ignored in the command input or just space and tab
+	while (*line == ' ' || *line == '\t')
 		line++;
 	// TODO: add parsing validation such as opening and closing quotes, single &...
-	if (!parsing_validation)
-		return (NULL);
 	return (get_commands(line));
 }
