--- srcs/parsing/validation/validation.c
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   validation.c                                       :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: oaarsse <oaarsse@student.42.fr>            +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/11 13:32:19 by oaarsse           #+#    #+#             */
-/*   Updated: 2022/10/11 13:38:51 by oaarsse          ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "common.h"
-#include "parsing.h"
-
-// TODO: add parsing validation such as opening and closing quotes, single &...
-// prints the parsing error too and returns false
-t_bool	parsing_validation(char *line)
-{
-	return (TRUE);
-}
